Move HydraBuilderfactory
